+++ cs/src/tables/keccak_precompile_related.rs
@@ -64,7 +64,7 @@ pub fn create_keccak_permutation_indices_table<F: PrimeField>(id: u32) -> Lookup
 
             let control = control_with_exe & 0b0111_1111_1111;
             let exe = (control_with_exe >> 11) == 1;
-            let precompile = control as u32 & 0b111;
+            let precompile = control & 0b111;
             let iter = (control as usize >> 3) & 0b111;
             let round = control as usize >> 6;
 
@@ -262,7 +262,7 @@ pub fn create_rotl_table<F: PrimeField>(id: u32) -> LookupTable<F> {
             debug_assert!(input < (1 << 20));
 
             let word_u16 = input as u16;
-            let rot_const = (input >> 16) as u32;
+            let rot_const = input >> 16;
 
             let (left, right) = (
                 word_u16.unbounded_shr(16 - rot_const),
